Added a connection create and destroy handler
For several cases it is usefull to run special tasks if a connection is
created or destroyed.
With
api.connections.createHandlers.push(connectionDestroyed);
you can add a handler thats called if a new connection is created and
with api.connections.destroyHandlers.push(connectionCreated);
you will be informed if a connection is destroyed.
You can use these e.g. to register a interval handler that is checking
or notifieng a backend or frontend if a user joins or leaves.