Fix for lack of javascript folder under public
ActionHero likes to change actionheroClient.js/.min.js so I elected to not track them as they'd be system-specific anyway. This resulted in the javascript folder not existing on other systems, which caused the tests to fail. This is a simple fix to create the javascript folder if it hasn't been.