Fix loading a locked item
An item is loaded from the cache and is eventually old copy, because first is loaded and then is checked for locking. In such a case the locker worker may apply changes to the item during the lock, but we don't receive a fresh copy for the item. This fix is checking for lock first and then tries to load the item.